[Map] extract out renderers into their own files for the large ones. extract out some of the hardcoded positions used on the map and split up the large initialization methods.
